Corriger la fin de suivi pour SSA
En cas de cloture d'une fiche, si la structure est la dernière on met la
fiche en fin de suivi. Pour cela on déplace la méthode du modèle
Evenement de SV dans le mixins qui gère la fin de suivi et les messages.
Ajout d'un test spécifique a cette situation.

Fixes SEVES-59
