+++ ssa/tests/test_evenement_produit_details_actions.py
@@ -39,6 +39,25 @@ def test_can_cloturer_evenement_produit(live_server, page: Page, mocked_authenti
     expect(page.get_by_text(f"L'événement n°{evenement.numero} a bien été clôturé.")).to_be_visible()
 
 
+def test_can_cloturer_evenement_produit_if_last_remaining_structure(
+    live_server, page: Page, mocked_authentification_user
+):
+    ac_structure = Structure.objects.create(niveau1=AC_STRUCTURE, niveau2=MUS_STRUCTURE, libelle=MUS_STRUCTURE)
+    contact = Contact.objects.create(structure=ac_structure)
+    evenement = EvenementProduitFactory(etat=EvenementProduit.Etat.EN_COURS, createur=ac_structure)
+    mocked_authentification_user.agent.structure = ac_structure
+    evenement.contacts.add(contact)
+
+    details_page = EvenementProduitDetailsPage(page, live_server.url)
+    details_page.navigate(evenement)
+    details_page.cloturer()
+
+    evenement.refresh_from_db()
+    assert evenement.etat == EvenementProduit.Etat.CLOTURE
+    expect(page.get_by_text("Clôturé", exact=True)).to_be_visible()
+    expect(page.get_by_text(f"L'événement n°{evenement.numero} a bien été clôturé.")).to_be_visible()
+
+
 def test_can_publish_evenement_produit(live_server, page: Page, mocked_authentification_user):
     evenement = EvenementProduitFactory(etat=EvenementProduit.Etat.BROUILLON)
